Refactor components to game/components directory
