+++ src/game/components/Pieces/Cube.jsx
