+++ src/game/components/Pieces/Pyramid.jsx
