+++ src/game/components/Pieces/Sphere.jsx
