+++ src/game/components/Pieces/Tetrahedron.jsx
